pyproject.toml
[project]
name = "BBRL2"
dynamic = ["version"]
authors = [
    { name = "Ludovic Denoyer" },
    { name = "Alfredo de la Fuente"},
    { name = "Song Duong"},
    { name = "Jean-Baptiste Gaya"},
    { name = "Pierre-Alexandre Kamienny"},
    { name = "Daniel H. Thompson"},
    { name = "Olivier Sigaud", email = "olivier.sigaud@isir.upmc.fr" },
    { name = "Mathis Koroglu", email = "mathis.koroglu@etu.sorbonne-universite.fr" }
]
maintainers = [
    { name = "Olivier Sigaud", email = "olivier.sigaud@isir.upmc.fr" },
]

description = "BBRL2 is a Flexible and Simple Library for Reinforcement Learning Agents"
keywords = ["sequential learning", "reinforcement learning"]

readme = "README.md"
license = { file = "LICENSE" }

requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "torchvision",
    "gymnasium",
    "hydra-core"
]

[project.optional-dependencies]
logger = ["tensorboard", "wandb", "pandas", "tqdm"]
plot = ["matplotlib", "seaborn"]
format = ["black", "pre-commit"]
all = ["tensorboard", "wandb", "pandas", "tqdm", "matplotlib", "seaborn", "black"]


[project.urls]
repository = "https://github.com/osigaud/bbrl2"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["bbrl*"]
namespaces = true

[tool.setuptools_scm]
write_to = "bbrl2/_version.py"

[tool.black]
line-length = 88
target-version = ['py37']

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
